Updated wvdial version to 1.40.
Added optional support for the liblockdev library to WvLockFile, and made
rules.mk able to merge .a files together (for now, so that liblockdev.a
could be linked automatically, but later, so we can easily build a
libwvstreams.a from all the various libraries).

More virtual function fixes in WvModem.

wvdialconf now checks /etc/ppp/options for common errors, and sends
people to the FAQ, then to wvdial-list@worldvisions.ca rather than asking
people to personally report problems to wvdial@worldvisions.ca


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@147 6619c942-59fa-0310-bf43-a07aaa81d9ac