Updated LDAP server to use a more standards-compliant layout.  It also now works with pam_ldap, so you can set up a debian system to authenticate against a Net Integrator's LDAP server.  Ooh.
Also made it so you don't need to specify any base dn when you configure an
ldap client.  Instead, the ldap server sends a referral to the right one.
At least, you can say goodbye to "o=net-itech.com".  That base dn was
totally nonstandard anyway - the correct one is "dc=net-itech,dc=com", which
we now use... but you don't have to type it anywhere so that's okay.

Also made NFS send a SIGHUP to mountd when it reconfigures /etc/exports.


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@293 6619c942-59fa-0310-bf43-a07aaa81d9ac