r3_70: big rearrangement of certificate management stuff: some web browsers get seriously confused if apache and webconfig are using different certificates (sigh).
Fixed a few other bugs in WvApache, including the tendency to not start
either the plain or ssl versions if one of them wasn't enabled at startup.
Made imap/pop servers not kill connections just to restart the listener (I
think).  Made the WebConfig <title> tag actually include the hostname, so if
you have lots of them open you can see what they are.  Fixed a bug in WvX509
certificate dumping - sometimes extra junk would be added to the end of the
file.  Changed the order of encode_hostname_as_DN so it puts the cn= part
last, rather than first (which seems to be slightly more common).


git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/r3_70@810 6619c942-59fa-0310-bf43-a07aaa81d9ac