Changed UniConfIniFile so that [section names] with equal signs in the name don't need to be tcl-escaped.
Massively restructed WvLdapClient.  It can only do read (search) queries,
but it can do lots of them in the background and does sane things with
WvStreamList and select().  It dumps the query results into a UniConf
branch, which is... interesting.  I sure wish we had a KDE UniConf
viewer/editor!  (hint hint...)


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@1215 6619c942-59fa-0310-bf43-a07aaa81d9ac