WvStreams makefile changes, part of the "The Miracle Diet Makefile" plan.
Added an "objects" function that takes directories and expands to the
name of the object files that should be in there (takes all the .cc
and .c files and turn them into .o). Removed a bunch of vars.mk files
by replacing them with uses of the "objects" function higher up.

Added a VERBOSE variable that is the moral equivalent of what DONT_LIE did,
so you get get the details when something goes wrong.

Added an implicit target for going from a .cc to an executable.

Including your subdirs vars.mk and rules.mk is now done at the top level,
so you don't need to do it anymore.

Switched from using funky variable variable names (like "libwvutils.so-LIBS")
to using target-local variables.

Fixed a number of mostly aesthetic buglets.


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@1364 6619c942-59fa-0310-bf43-a07aaa81d9ac