Created a shell script to test the UniConfDaemon more thoroughly. Currently looks at all files of the format:  daemontest*.txt and passes them to nc to send to the daemon.
Results are then recorded in res_daemontest*.txt files, and diffed with expected
results.  If there is a difference, a less session is brought up with the results
of the diff.

There is a help feature to the script, accessible by daemontest.sh -h

To add new tests, simply create a new daemontest*.txt file, and a new
real_res_daemontest*.txt file containing the expected results of the
operations.


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@1384 6619c942-59fa-0310-bf43-a07aaa81d9ac