Fixed a bug in wvresolver which would make name resolution fail if you were only selecting for write, which 'uni set' tends to do.  Fixed a related bug in WvStream::flush_outbuf() where it would select() for nonzero time with an empty fdset because WvStream::isok() is always true, the outbuf was nonempty, and the derived class was in the process of being destroyed so its own isok() (which was false) was no longer accessible.  Oops!
Added a new "xset" command to uni, which reads the value from stdin.  "uni
xset /etc/profile </etc/profile".  As long as the file contains no nuls,
the UniConf value is byte-for-byte the same as the file contents.  This may
seem silly *now*, but it'll be useful with the upcoming IMAP backend :)


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@2448 6619c942-59fa-0310-bf43-a07aaa81d9ac