Found a better way to serialize the standard int types. Changed wvgdbmhash to tell gdbm to not fsync() after every write, which makes it go about a bazillion times faster.  Split wvgdbmlist into its own files.  FIXME: It's not very optimized, and everything is inlined.
Made WvTape optionally use WvGdbmList for its WvTapeFileList object.  This
makes the initial indexing go about 10 times slower (if all the files are
already in cache) but drops the memory usage of wvtar from 25 megs and
rising to a flat 2.5 megs in my test.  But because GdbmList doesn't provide
a Sorter, this isn't actually workable quite yet.  Maybe a btree would make
more sense than sorting an on-disk list...


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@2528 6619c942-59fa-0310-bf43-a07aaa81d9ac