Rewrote WvTimeStream to be more like it used to be before the alarm() reimplementation (which broke the interface contract of the class), with the difference that it is now implemented using the wvtimeutils.h primitives, WvTimeStream::pre_select will actually adjust the si.msec_timeout (so it is now safe to do select(-1)) and I didn't bring back the max_backlog feature, which didn't seem to be used anywhere.
Also fixed a number of bugs in weaver and lcd where WvTimeStream::tick was
being called directly (it is an error since the moment that it was getting
called automatically in WvTimeStream::execute), with a horrible hack in
WvLCD, which is an horrible hack in itself (two wrongs *do* make a right!).

Ever noticed how the backup countdown was taking more than a second per
second? :-)


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@2556 6619c942-59fa-0310-bf43-a07aaa81d9ac