Enabled some standard conformance options to get rid of a few warnings in VS.
Added a zlib 1.1.4 binary to cvs. It's massively spiffy so we'll probably never have to upgrade it.

Upgrade the zlib dll in the system32 directory in the installer but install the openssl dlls with the application (since the openssl team breaks binary compatibility from time to time).

Fixed the line endings in some files. John is now gone and I don't use crummy cygwin so I even have a running chance in winning this battle.


git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@2761 6619c942-59fa-0310-bf43-a07aaa81d9ac