r3_75: Hold the pointer (*wvauthd) to the global WvAuthDaemon within WvConf.  This is an ugly hack, but it allows WvConf easy access to the daemon without having to change the interface of WvConf.
I had considered making WvConf contain virtual check_passwd()
functions; but this causes the linker to complain bitterly that not
all symbols are found.  We don't ship wvconf.cc with implementations
of get_passwd(), set_passwd(), check_passwd() or user_exists(); so
this is not a good solution.

WvConf itself does not manage *wvauthd but WvAuthDaemonSvc does.  This
service is a friend of WvConf, which allows this to happen in a
protected way.

BUGZID: 1794


git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/r3_75@3099 6619c942-59fa-0310-bf43-a07aaa81d9ac