carambola: merged recent changes from r4_1+ns3, except the binary kernel and modules.  Someone will have to rebuild the kernel eventually, but I expect they'll do that pretty soon anyway.  (That means we're currently missing the watchdog-timeout patch and the usb-serial-write-oops-fix patch.)
The new unitransactiongen unit test revealed bugs in carambola's
UniTempGen::exists() (which I fixed by just using the UniDefGen default
implementation; there's no advantage to special casing this) and
UniUnwrapGen when used with trailing-slash UniConf keys, the behaviour for
which changed in carambola.  These two bugs, in turn, caused odd behaviour
in UniTransactionGen, so we're rather lucky that its unit test caught it.


git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/carambola@9685 6619c942-59fa-0310-bf43-a07aaa81d9ac