carambola: wrapped the entire UniConf tree in the main weaver process in a UniTransactionGen.  This makes get() return the new value immediately after set(), which is the kind of semantics that everything in the weaver process is expecting.  It probably fixes the bug where PDR says "That version is not on the system" when trying to do "wv upd switch" from carambola or zanykid; there wasn't enough time for the set() in WvSoftUpdate::scan_missing_images() to propagate back, so get() said the new image didn't exist.
Note that it's strictly correct for UniCacheGen to wait for a notification
after set() before get() returns the new value, because the change really
might be rejected by the inner generator.  UniTransactionGen seems to add
exactly the desired semantics, but hopefully it won't break anything.  I did
have to make UniTransactionGen actually remember to call the inner
commit()/refresh() functions, which it forgot to do before.


git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/carambola@9849 6619c942-59fa-0310-bf43-a07aaa81d9ac