Fixed a bug in wvtcl_getword() that was improperly treating \ characters within braces as the start of an escape sequence.  This means that a \ at the end of a value that is enclosed in {} (i.e. "{\}") will cause the } to be missed, and wvtcl_getword() will never see balanced braces, and thus always returns 0.
This was found through idbd setting a key in UniConf, causing UniConf to stop accepting commands over its connection.  As such, UniConf and everything that uses it will need extensive testing to be sure this hasn't broken anything.  I've added a unit test for idbd, but other areas may be affected as well.

BUGZID:17210


git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/kid_c@10880 6619c942-59fa-0310-bf43-a07aaa81d9ac