Added WvDBusMsg tests and added more ways to get at data from WvDBusMsg::Iter.  Improved WvDBusConn::add_callback() so it handles prioritizing your callbacks: you can now have "last chance" callbacks that handle messages only if you didn't handle them internally.  With this feature, WvDBusServer can now do its proxying without overriding WvDBusConn::filter_func, which means we actually don't need to override WvDBusConn at all.  So I got rid of WvDBusServConn, and got rid of WvDBusServerPrivate while I was there since it was just causing confusion.
The new WvDBusServer code uses only the new, simple add_callback(), not the
WvDBusListener stuff at all.  It turns out this actually makes it shorter
*and* easier to read.

Changed a lot of the message logging stuff in the server to refer to
connections by their assigned uniquename, instead of their registered name.
To help with that, started using a serial number instead of a random number
in response to the Hello() request.

While I was there, I made WvStream auto-unregister itself from
IWvStreamList::globallist during its destructor.  Why not?




git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/uniservice@11706 6619c942-59fa-0310-bf43-a07aaa81d9ac