wvtesthelper: work even if WVSTREAMS_SRC isn't set.
WvMap: WvMap's API was kind of annoying for maps where the data object
was a pointer, eg. WvMap<WvString,MyObject*> map.  In that cast,
map.find(key) would return a MyObject**, which was useless, and
map[key] would return a MyObject* but would fail an assertion if key
wasn't found, even though it could have just returned NULL (as
WvHashTable does).

Changed it to work more ilke WvHashTable: map.find(key) and map[key]
are now identical, and always return a simple TData (which is a
pointer).

*That* caused confusion, however, for non-pointer data types:
WvMap<WvString,int>.  Here, TData is not a pointer, so you can't
normally return NULL.  To deal with that, we have find() and [] return
objects of type TData* (not TData), but *only* if TData isn't already a
pointer.  This makes the syntax perhaps slightly surprising, but always
what you actually want.

Fixed up a few places where this makes the syntax incompatible
with what it used to be.  (In all cases, the code is now simpler.)



git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/uniservice@11708 6619c942-59fa-0310-bf43-a07aaa81d9ac