Rewrote WvDBus to do all the connection/authorization/etc stuff by itself. The advantage of this is we can now be fully asynchronous, and with only a bit more work, we should be able to wrap connections in SSL, etc. However, it still fails a few unit tests - I just need to go to bed now and I didn't want to not check it in first.
As part of this, introduced a new IWvListener/WvListener class like I've
been meaning to do for a while, so that WvUnixListener and WvTCPListener
can have a shared interface (and eventually monikers) for accepting new
sockets.  I only translated WvTCPListener for now - WvUnixListener
should be fairly straightforward to do next.

To do *that* properly, I needed to reintroduce the IObject* parameters to
wvcreate().  I'm not sure who took them away, but I said I needed them,
darn it, and I did.  The majority of code changes are related to that.


git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/uniservice@11722 6619c942-59fa-0310-bf43-a07aaa81d9ac