Fix WvDBus unit tests.
Previously, we were passing a unix: moniker to WvDBusServer.  However,
WvDBusServer didn't actually used to take a moniker, it would just pass
the string directly to a WvTCPListener, which interpreted it as an IP
address, and defaulted to 0.0.0.0 when given an
obviously-not-an-IP-address string like a unix: moniker.
WvDBusServer::listen() now takes a proper moniker, but this exposes the
fact that WvDBusServer::get_addr() only works for tcp: connections.  Oops.

So the quick fix is to just specify a tcp:0.0.0.0 moniker off the bat,
as it's what we've always been using anyway.