The callback should be called on a stream even if !isok().
This set of bugs was causing trouble with WvDelayedCallback, and probably
other things.  It probably was causing a rare crash bug in WvHttpPool too.

Basically, we would never call pre_select() or post_select() if !isok(),
which was unnecessarily "too safe."  It's their job to decide if they
don't want to do stuff because they're not isok(), and they were doing
pretty much fine at that job, so removing the checks was harmless.

However, removing the checks changed the behaviour of the
IWvStreamList::sure_thing list subtly, revealing that we forgot to addRef()
a stream upon adding it to that list.  In rare conditions where a stream
would auto-remove itself from a streamlist in the middle of a select()
sequence, the stream would be deleted but still in sure_thing, causing weird
bugs.

So now we addRef() before adding a stream to sure_thing, where we now set
auto_free=true so it will release() automatically.  We also have to addRef()
when taking it *out* of sure_thing to call it (since we've always
deliberately removed the stream from sure_thing *before* using it to avoid
other weird race conditions) and release() it when we're done.

Added new unit tests that verify that alarm() will still work even after
a stream is close()d.  But only once, of course, since the stream is dropped
out of the list after that.