Fixed wvdelayedcallback unit test.
Based on recent changes to WvDelayedCallback, deleting a
wvdelayedcallback doesn't actually kill the stream inside it.  As such,
the test where a wvdelayedcallback is killed, globallist.runonce() is
executed, and the value of a variable updated by the contained stream
was expected to remain unchanged, will now actually change.  Unit test
updated.