Wvdelayedcallback no longer resets callback of stream within.
When we kill a wvdelayedcallback, this no longer resets the callback
of the stream contained within to (0), had it been triggered.  We want
this behaviour so that we can assign delayedcallbacks to the
'setclosecallback' function, which immediately deletes the callback
after triggering (which, previously, meant that the stream contained
in a wvdelayedcallback would never go off if the callback was assigned
to setclosecallback).