WvDBusServer no longer causes segfault when connection dies.
WvDBusServer had a wonderful little glitch in it.  If a connection
died, a callback was invoked to remove it from the linked list of
all connections in the server.  This is fine, except one of the
places a connection could die was while we were in a 'broadcast'
loop running an iterator over all the connections.  Once we
removed a connection from the list in this loop, the poop hit the
fan, of course, as the iterator we were on was no longer meaningful.

Solution?  WvDelayedCallback for removal!  Some other changes had to
be made, to refcount the connection (so that it doesn't get deleted
until we tell it to in the callback), as well as for the DBusServer
object itself (so that the delayed callback never runs on a deleted
DBusServer, trying to remove connections from its instance variables).