Merged most of the tiny crap Makefile fragments into Makefile.
But left a few behind in crap*, to be cleaned up later.  Basically those are
the unit tests, which don't work for the moment.

Remove weird support for building xplc as part of wvstreams.

Remove old msvc build stuff.  I'm not sure anybody *ever* used it, but
I'm sure it's broken now, and it would have to be totally redone anyway.

I *really* don't want to know why Makefile was setting VPATH.