Atomically create/update files generated by gen-cc.
Running make -j2 would sometimes cause gen-cc to be run twice at once.
Since gen-cc removed the existing target file before rewriting it, the
second gen-cc instance would remove the file while the Makefiles were
trying to use it.  Now we write to a temp file, and atomically move it
on top of the target file.