Added 'fingerprint' grabbing to WvX509.
Unbeknownst to me until today, openssl has a way of grabbing
the 'fingerprint' of an SSL certificate.  Since this is a
fairly good way of figuring out whether we have the certificate
we're looking for, it's handy to get this value out of
WvX509.  Done, via get_fingerprint(mode), where mode will
determine whether we get a SHA-1 or MD5 cert.