Added sslcert moniker to WvSSLStream.
Now, as a client needing to pass a certificate into SSL, I can do so,
via this handy moniker which takes a WvTclString as an argument.
Given an SSL certificate in PEM format, and an RSA key, as well as a
connection moniker, we create an SSL stream which uses a WvX509Mgr
with the cert and RSA key, connected to the given moniker.