Modified WvMonikerRegistry to allow multiple instances of a moniker.
Allowed moniker overriding by changing the backing store of a
WvMonikerRegistry to a linked-list, instead of a hashtable, and
doing linear searches for a moniker, prepending new (overriding)
ones to the beginning.  It's not /perfect/ in that you can do
strange things and remove incorrect monikers (ie. not the ones
that you actually want) this way, but you'd have to be doing very
silly stuff.