Fixed static linking with WvMonikers and Mingw32.
Mingw32 had the awesome property (with version 4.2.1) that it optimised
out WV_LINK_TO lines, as the anonymous namespace contained therein was
no longer considered exported (gcc 4.2.x), and was unused in the local
file.  This has been changed into a static function, which, thanks to
the wonders of gcc, is /not/ optimised out until after designating that
we should link in other .o modules containing the monikers we want.

The downside is that we use the GCC 'unused' attribute to prevent
compiler warnings, and so this isn't portable to Windows.  If anyonee
wants to fix WvStreams compilation on Windows, you should look at this!