Try another fix for wvlinkerhack and static libraries.
This time we use volatile ints instead of normal ints.  Theoretically, the
read of the WV_LINK() volatile int should not be allowed to be optimized
away, nor should the store to the WV_LINK_TO() volatile int.  Sadly, this
will generate actual init-time code in the program, unlike the old system
that just generated a symbol reference, but any compiler that optimizes it
away is broken.