Add a global object in wvx509.cc to prevent SSL errors from getting unloaded.
I suspect this is only needed in OpenSSL 0.9.7, so protect it with an ifdef.
This loads SSL error messages at the start of the program, which is kind of
lame since the program might not even need them.  But it's the only safe way
to delay *unloading* the messages until the end, assuming we still want
valgrind's memory leak checker to work in unit tests.