Flush connections for all DBus server tests, not just the last one.
WvDBusServer objects are refcounted for each open connection.  So they
don't necessarily go away immediately when you delete them.  This can
cause Valgrind to complain about memory leaks in the unit tests, and/or
the open file descriptor check to fail.

The reason these things didn't normally happen is that, as it happened,
the last test in the file had code to run the globallist a few times to
clear everything out.  Move this code to the TestDBusServer destructor,
so now all the individual tests work as expected.