WvDBusServer: remove serial_to_conn map.
That's weird; it turns out serial_to_conn is completely unnecessary, because
the destination of a reply message is always set to the person you're
replying to anyway.  I'm not sure why I thought otherwise.

Remove the silly mapping function.  As a bonus, this gets rid of ridiculous
problems that can be caused when different clients produce messages with the
same serial number, which is not only possible, but common.