wvgzip: use Z_BEST_SPEED instead of Z_DEFAULT_COMPRESSION.
Since this is WvStreams, we can assume people using WvGzip will want to use
it for network stuff, and in that case, speed is more important than
compression ratio.  The speed difference this makes is pretty huge.

It might be better to add an option to let you set it, but it didn't seem
worth the trouble right now since hardly anyone uses WvGzip anyhow.