Makefile:  Massively clean-up argp inclusion in wvstreams.
Hacks are a thing of the past with this new clean-up.  Pull in the
libargp.list file for libwvstatic.a, link the libargp.a library for
libwvutils.so.  No more @echo -n's, strange-case variables, or .PHONY
targets which look like they shouldn't be phony.

Bonus:  works with -j2, though there are still other parts of the
Makefile which don't.