Remove the NULL-related check in wvstring.t.cc, and the warning it caused.
WvString(NULL) actually produces a wvstring containing "0", because in C++
NULL is defined as 0, not (void *)0, and thus the pointer-based WvString
constructor doesn't run.  It has always been like that, since some version
of gcc where the correct C++ behaviour was added.  The unit test just
checked that this was in fact the case, in case it ever changed.

Signs are that it might, in fact, change when C++0x comes out.  In any case,
checking the current behaviour is somewhat pointless, because anyone who
*relies* on the current behaviour is obviously crazy anyway, and it might
indeed change - for the better - when C++ changes.

So let's git of this warning.  I've hated it for years anyway.