"wvtestmain ''" actually still runs all the tests, so skip the warning.
wvtest.cc prints a message "WARNING: WvTest: only ran tests starting
with specifed prefix(es)" if you give it any parameters restricting the
tests to be run.  But 'make runtests' actually sometimes provides an
existing but empty parameter, which doesn't restrict anything because *any*
test starts with the empty string.  So suppress the warning in that case;
it's the desired behaviour.