Oops, some Makefile rules depended on "WVSTREAMS=."
If we had a dependency on $(WVSTREAMS_LIB)/libwvdbus.so, for example, it
wouldn't be satisfied by a rule for just "libwvdbus.so", but it works fine
if $(WVSTREAMS_LIB) is ".".  So change it back.

To fix the LD_LIBRARY_PATH bug from earlier, just set it to `pwd` instead of
WVSTREAMS_LIB.