wvtest.cc: consider it an error if 0 tests were run.
When there's a linker problem of some sort, you might end up linking in
wvtestmain without any actual test programs, and 'make test' will succeed
because no tests failed.  But that's obviously not what you really wanted,
because no tests *passed* either.  We add a check at the end of run_all() to
ensure that at least one test was run; if not, we return a failure code.

versaplexd had this problem for a while, and I'd rather not ever see it
again.