A bunch of fixes to do a few things:
First of which is to fix the distclean
target to actually clean up argp stuff properly.

Second is to add proper detection of AR for argp.

Third is to begin Solaris port, and get it to the point where at least it
compiles all of the code, and creates libwvstatic.a.

Signed-off-by: William Lachance <wrlach@gmail.com>