Makefile: add 'autodep-test' target for testing autodependencies.
People (including me, sometimes) keep on breaking header autodependencies
and not noticing, and it wastes a lot of time.  Now the Makefile has a
self-test to ensure the dependencies are (at least basically) working
correctly.

'make test' now depends on the autodep-test target so that you can't forget
to run it.