WvStream: use a different workaround if the streamlist is empty on win32.
On win32, ::select() won't delay *at all* if you give it an empty fd_set,
which is lame.  To work around this, we were apparently trying to create a
temporary socket and select on *that* instead, but this didn't work very
well - maybe the problem is only in wine, but it's still non-ideal.

Instead, let's just Sleep() when we know the fd_set is empty.