wvwin32task: don't get confused if WvCont is used on multiple threads.
You probaby shouldn't try to WvCont *between* threads, but previously,
WvCont would *only* work on the very first thread where you created your
very first WvCont.  In most apps, that's actually probably fine.  But it's a
bit of a pain for unit tests, so improve things a bit.  Now a given WvCont
should work on the thread where it was first created.