WvError: add a bunch of constructors and a handy setter.
It turns out that WvError objects are kind of handy to copy around and use
as return values, which is something we never used to do in the "old"
WvStreams, but which random applications can actually benefit from.

With that in mind, add a bunch of WvError() constructors so you can do
things like

   return WvError("myfunction", my_func(...));

(that constructs a new WvError and if my_func() returns a WvError that's
!isok(), it turns the error message into "myfunction: %s" where %s is the
original error message.  If the original error *is* isok(), it just leaves
the new error entirely unset, as you'd want.