WvFile: fix serious bugs in pre_select/post_select.
In Unix, we were changing si.wants without changing it back.  WvIStreamList
changes it back for self protection, but I'm pretty sure we shouldn't leave
it lying around.

In Win32, selecting on a WvFile doesn't work at all, and in fact gives an
"invalid argument" error that causes the file to close, so getline() doesn't
work on a WvFile.  That's kind of nasty.  Work around it by just always
assuming a WvFile is readable/writable.  (In Unix, this isn't always true,
because a WvFile might be a device node.  But in Windows, that doesn't
really happen.)