include/wvhttppool.h:  Fixed WvBufUrlStreams with no error but no data.
So, problem:  You've got this WvBufUrlStream you just got back from a
request.  It's got '200' as the return code, and the headers specify
61 bytes of data, but you get none. What happened?

Turned out, if the *first* time the WvUrlRequest containing this
stream went through a WvHttpRequest, that one timed out (after making
a connection and sending data, ie. if the CGI on the server was taking
too long), it called seterr(...) on the underlying WvBufUrlStream,
which called dead in WvBufStream, which killed the stream for further
writing.  That's bollocks, because WvHttpPool can actually retry the
stream, and if the retry succeeds, you have no data.

We now override the seterr(int) and geterr() methods for
WvBufUrlStream so that it's unaffected by such things, meanwhile,
if the return code the user gets is not 200, they can geterr() to see
if there is any additional information provided to us by WvStreams.