WvHttpStream: don't delay if the last request is a pipeline check.
There's some funny code (which was presumably added for a real reason) that
waits for a possible HTTP response body even as the answer to a HEAD
request.  This may have been necessary in HTTP/1.0 or something, but it's
crazy to do this when using HTTP/1.1 with persistent connections, as you
won't know whether the body is over until the *next* response comes in - and
if you don't have any more requests, then there won't *be* a next response.

Sure enough, we just ran into this problem.  In honour of the original code,
we still read the body if it's an HTTP/1.0 server or we get Connection:
close.