Remove MacOS-specific readline hacks to wsd.
Essentially, wsd.cc was commented out entirely on MacOS, then we decided not
to build it, but 'make install' would try to build and install it anyway, and of
course fail because _main() didn't exist, having been commented out.

The correct fix is to actually look for the parts of libreadline we actually
want, and if they're not there, mark readline as not being available.  Then
we don't need any special MacOS hacks.