Improving (but not quite fixing) wvhttpstream bug
If a wvhttpstream is too fast to connect (such as to a server running on
the same machine), the connection may be established before putstream is
done being read, in which case for some reason we don't get another
callback, and the request doesn't get sent.  We can at least reduce the
chances of this happening.