Revert OpenSSL 1.1 compat changes.
The WvOCSP change requires at least OpenSSL 1.1.1, which isn't yet in
debian-stable (they have 1.1.0f).  1.1.0f seems to be incompatible with
both the old and new code, but at least 1.0.2l is still available and
works with the old-style code, for now.

Still seeing some test failures with WvX509 and WvSSLStream even after
the revert, but that's probably unrelated.

This reverts commit 5506a74e1b033776ad441b4554716cdcfa88fe03.