Remove unnecessary && and 'exec >&2' paranoia from .do files.
I had forgotten that apenwarr/redo always calls with 'sh -e', so sh
scripts will abort after the first error anyway.  And redo also gives a
useful error message if you write to stdout *and* $3, so we don't have
to be so defensive about redirecting stdout all the time.

This makes the scripts a little more readable, and thus more suitable
to use as redo examples.