Remove hardcoded wvports-related paths in Windows builds.
Instead, you should pass --with-zlib=$PWD/.... etc directly, to specify
paths to any dependencies you might have built elsewhere.  This is what
the wvbuild project was already doing when building wvstreams, but the
presents of the wvports-related specific paths was hiding a bug in our
implementation of the --with_zlib parameter.