Refactor of delegate.od
`delegate.od` is the third-largest [redo](https://github.com/apenwarr/redo)-related file in the repository, and contains considerably more complexity(/generality) than `_config.od.in` or `libwvbase.list.do`. Following some redo mailing list discussion, this is my attempt to decompose it into independent functions.

Note: I have not attempted to run this code, or otherwise set up an environment capable of building wvstreams. Mostly this is a draft, with best-effort preservation of semantics, submitted to @apenwarr for regularization to local coding style and/or commentary on the ergonomics of *redo* as a tool :)