configure: improved way of detecting if redo is installed.
Previously, we'd give the same error no matter why redo returned a
non-zero exit code, which could be for a variety of reasons, but we
redirected output to stderr, so redo couldn't report which one.
Instead, detect whether there is a runnable 'redo' command at all, and
only print our warning message if not.

Reported-by: Nathan Myers